Created new cs files for ReferenceFixer and SearchPathsAssemblyResolver
