+++ src/generator.csproj
@@ -122,6 +122,8 @@
     <Compile Include="..\src\bgen\Nomenclator.cs" />
     <Compile Include="..\src\bgen\NullabilityInfoContext.cs" />
     <Compile Include="..\src\bgen\PlatformNameExtensions.cs" />
+    <Compile Include="..\src\bgen\ReferenceFixer.cs" />
+    <Compile Include="..\src\bgen\SearchPathsAssemblyResolver.cs" />
     <Compile Include="..\src\bgen\StringExtensions.cs" />
     <Compile Include="..\src\bgen\ThreadCheck.cs" />
     <Compile Include="..\src\bgen\TrampolineInfo.cs" />
